% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_pow.R
\name{sample_pow}
\alias{sample_pow}
\title{Creates a plot for the sample size and the power based on the standard deviation.}
\usage{
sample_pow(sd_ber = T, delta = 0, Delta, sd, test = 1, alpha = 0.05, beta = 0.2,
           prop = c(0.5, 0.7), adj = F,  regel = F, nbound = 500,
           fix_sim = c("fix", "sim"), simu = 10000, create = "plot")
}
\arguments{
\item{sd_ber}{Sequence of numbers. Interval of the actual standard deviation in the data.
With regard to this interval, the sample size and the power are displayed.
If no interval with more than two numbers but an upper limit is defined (see nbound),
a reasonable range is calculated automatically.}

\item{delta}{Number. Expectation difference of two samples.
If you select a Test for superiority/ difference then select 'delta = 0'.
Only if you select a Test for non-inferiority you can select 'delta != 0'.
Attention: If you chose 'test = 1' and 'delta != 0', the test for non-inferiority will automatically
be applied.
If not specified, delta is set to 0.}

\item{Delta}{Number. Relevant difference of expected values in the alternative hypothesis.}

\item{sd}{Number. Assumed standard deviation of the data.
Used to calculate the originally planned sample size.}

\item{test}{Number. What type of hypothesis test should be performed, one-sided (Superiority/ Non-Inferiority test)
or two-sided (Test for difference).
One-sided (test = 1): Superiortity H0: mu_x - mu_y <= 0 vs. H1: mu_x - mu_y >0
                      Non-Inferiority H0: mu_x - mu_y >= delta vs. H1: mu_x - mu_y < delta
Tweo-sided (test = 2): Difference H0: |mu_x - mu_y| = 0 vs. H1: mu_x -  mu_y != 0
Attention: Choice of delta. (see delta)
If not specified, the one-Sided Test (Superiority/ Non-Inferiority Test) is used.}

\item{alpha}{Number. Desired alpha-level of the test.
If not specified, alpha is set to 0.05.}

\item{beta}{Number. Acceptable beta error of the test.
If not specified, beta is set to 0.2.}

\item{prop}{Number/ vector of numbers.
Timing of the internal pilot study depending on the originally planned sample size.
If a vector is passed, all timings are displayed.\cr
A maximum of five different values are possible if you choose \code{create = "tab"}.}

\item{adj}{Logical. Should the one-sample variance, calculated in the internal pilot study, be adjusted?}

\item{regel}{Logical. Should the sample size adjustment rule be applied by Wittes and Brittain?}

\item{nbound}{Number. Upper limit of the sample size.
Attention: Only if you choose nbound can a suitable standard deviation range for the plots
be calculated automatically. If no nbound are defined then a standard deviation range must be chosen
(see sd_ber).}

\item{fix_sim}{Which design should be applied and presented?
"fixed": Design with fixed sample size,
"sim": Design with internal pilot study,
c("fix", "sim"): both design}

\item{simu}{Number. How many simulations should be performed?
If not specified, simu is set to 10000.}

\item{create}{"plot" if a plot is to be returned.\cr
"tab" if a table is to be returned.}
}
\value{
It returns a plot.
}
\description{
That is one of the main functions.
It draws a plot of the sample size and the power based on the standard deviation.
The sample size and power can be calculated for the design with a fixed sample size and / or
with an internal pilot study.
If desired, it may represent several timings for the internal pilot studies for comparison.
}
\seealso{
\code{\link{fix_sample_pow}}\cr
\code{\link{sim_sample_pow}}\cr
\link{ggplot2}\cr
\link{gridExtra}
}
\author{
Csilla van Lunteren
}
