% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow.R
\name{sim_p_calc}
\alias{sim_p_calc}
\title{Decision regarding test statistics (Design with internal pilot study)}
\usage{
sim_p_calc(n1, delta = 0, Delta, sd, test = 1, alpha = 0.05, beta = 0.2, N0, 
adj = F, regel = F, nbound = 500)
}
\arguments{
\item{n1}{Number. Size of the internal pilot study.}

\item{delta}{Number. Expectation difference of two samples.
If you select a Test for superiority/ difference then select 'delta = 0'.
Only if you select a Test for non-inferiority you can select 'delta != 0'.
Attention: If you chose 'test = 1' and 'delta != 0', the test for non-inferiority will automatically 
be applied.
If not specified, delta is set to 0.}

\item{Delta}{Number/ Vector of numbers. Relevant difference of expected values in the alternative hypothesis.}

\item{sd}{Number. Assumed standard deviation of the data. 
Used to calculate the originally planned number of cases.}

\item{test}{Number. What type of hypothesis test should be performed. One-sided (Superiority/ Non-Inferiority test)
or two-sided (Test for difference).
One-sided (test = 1): Superiortity H0: mu_x - mu_y <= 0 vs. H1: mu_x - mu_y >0
                      Non-Inferiority H0: mu_x - mu_y >= delta vs. H1: mu_x - mu_y < delta
Tweo-sided (test = 2): Difference H0: |mu_x - mu_y| = 0 vs. H1: mu_x -  mu_y != 0
Attention: Choice of delta. (see \code{delta})
If not specified, the one-Sided Test (Superiority/ Non-Inferiority Test) is used.}

\item{alpha}{Number. Desired alpha-level of the test.
If not specified, alpha is set to 0.05.}

\item{beta}{Number. Acceptable beta error of the test.
If not specified, beta is set to 0.2.}

\item{N0}{Number. Size of the originally planned sample size.}

\item{adj}{Logical. Should the one-sample variance, calculated in the internal pilot study, be adjusted?}

\item{regel}{Logical. Should the sample size adjustment rule be applied by Wittes and Brittain?}

\item{nbound}{Number. Upper limit of the sample size.
Attention: Only if you choose nbound can a suitable standard deviation range for the plots 
be calculated automatically. 
If no nbound are defined then a standard deviation range must be chosen (see \code{sd_ber}).}
}
\value{
This function returns a value if the simulation study rejects (1) or can not reject (0) the null hypothesis.
}
\description{
This is an auxiliary function of \code{\link{pow}}.
}
\seealso{
\code{\link{test_h0}}
}
\author{
Csilla van Lunteren
}
