% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_sample_pow.R
\name{fix_calc}
\alias{fix_calc}
\title{Decision regarding test statistics (Design with fix sample size)}
\usage{
fix_calc(N0, sd_ber, delta = 0, Delta, test = 1, alpha = 0.05, beta = 0.2)
}
\arguments{
\item{N0}{Number. Size of the originally planned sample size.}

\item{sd_ber}{Number. Actual standard deviation in the data.}

\item{delta}{Number. Expectation difference of two samples.
If you select a Test for superiority/ difference then select 'delta = 0'.
Only if you select a Test for non-inferiority you can select 'delta != 0'.
Attention: If you chose 'test = 1' and 'delta != 0', the test for non-inferiority will automatically 
be applied.
If not specified, delta is set to 0.}

\item{Delta}{Number. Relevant difference of expected values in the alternative hypothesis.}

\item{test}{Number. What type of hypothesis test should be performed, one-sided (Superiority/ Non-Inferiority test)
or two-sided (Test for difference).
One-sided (test = 1): Superiortity H0: mu_x - mu_y <= 0 vs. H1: mu_x - mu_y >0
                      Non-Inferiority H0: mu_x - mu_y >= delta vs. H1: mu_x - mu_y < delta
Two-sided (test = 2): Difference H0: |mu_x - mu_y| = 0 vs. H1: mu_x -  mu_y != 0
Attention: Choice of delta. (see \code{delta})
If not specified, the one-Sided Test (Superiority/ Non-Inferiority Test) is used.}

\item{alpha}{Number. Desired alpha-level of the test.
If not specified, alpha is set to 0.05.}

\item{beta}{Number. Acceptable beta error of the test.
If not specified, beta is set to 0.2.}
}
\value{
This function returns a value if the simulation study rejects (1) or can not reject (0) the null hypothesis.
}
\description{
This is an auxiliary function of \code{\link{fix_sample_pow}}.
}
\author{
Csilla van Lunteren 

 @seealso
\code{\link{test_h0}}
}
