% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_h0.R
\name{test_h0}
\alias{test_h0}
\title{Test a test statistic to see if the null hypothesis can be rejected.}
\usage{
test_h0(test = 1, T, n1, n2, alpha = 0.05, delta = 0)
}
\arguments{
\item{test}{Number. What type of hypothesis test should be performed. One-sided (Superiority/ Non-Inferiority test) (test = 1)
or two-sided (Test for difference) (test = 2).
One-sided (test = 1): Superiortity H0: mu_x - mu_y <= 0 vs. H1: mu_x - mu_y >0
                      Non-Inferiority H0: mu_x - mu_y >= delta vs. H1: mu_x - mu_y < delta
Tweo-sided (test = 2): Difference H0: |mu_x - mu_y| = 0 vs. H1: mu_x -  mu_y != 0
Attention: Choice of delta. (see delta)
If not specified, the one-Sided Test (Superiority/ Non-Inferiority Test) is used.}

\item{T}{Number. Value of the calculated test statistic.}

\item{n1}{Number. sample size of the first sample.}

\item{n2}{Number. sample size of the second sample.}

\item{alpha}{Number. Desired alpha-level of the test.
If not specified, alpha is set to 0.05.}

\item{delta}{Number. Expectation difference of two samples.
If you select a Test for superiority/ difference then select 'delta = 0'.
Only if you select a Test for non-inferiority you can select 'delta != 0'.
Attention: If you chose 'test = 1' and 'delta != 0', the test for non-inferiority will automatically 
be applied.
If not specified, delta is set to 0.}
}
\value{
The value 0 or 1 will be returned. 
0 means the null hypothesis can not be rejected to the level alpha.
1 means the null hypothesis is rejected to the level alpha.
}
\description{
This is an important secondary function. 
It checks whether the corresponding null hypothesis can be rejected by a calculated test statistic. 
Three cases of hypotheses can be considered: test for superiority, test for non-inferiority, test for difference.
}
\details{
The aim of this function is to simplify the application of test decisions in hypothesis tests.
}
\author{
Csilla van Lunteren
}
